Pass if/else expression VM tests
This involved a massive rewrite of the VM internal instructions structs.
We do not use a byte array, so implementing jump was not as simple as
incrementing the instruction pointer. Instead I made a new struct to hold
the instructions and the pointer. This also required me to rewrite `run`
with a `run_op` helper for pattern matching.